requirements.txt
PyQt5
PyQt5.QtWidgets
PyQt5.QtCore
google.generativeai
os
traceback